feat: support remote disconnects
